functions/manual_exercises/predictions_endpoint/requirements.txt
Flask
google-cloud-storage
db-dtypes
scikit-learn
pandas
joblib